feat: Add DEFAULT_PET_IMAGE constant and checkIfProvidedUrlIsImage utility function

+++ src/app/actions/actions.ts
@@ -1,16 +1,23 @@
 "use server";
+import { PetIdSchema } from "./../../lib/validations";
 
-import { PetEssentials, PetWithId } from "@/components/types/types";
+import { PetWithId } from "@/components/types/types";
 import prisma from "@/lib/db";
 import { sleeper } from "@/lib/utils";
+import { FormSchema } from "@/lib/validations";
 import { revalidatePath } from "next/cache";
 
-export async function AddPet(petDate: PetEssentials) {
+export async function AddPet(petDate: unknown) {
   await sleeper(2000);
+
+  const validatedPet = FormSchema.safeParse(petDate);
+  if (!validatedPet.success) {
+    return { error: validatedPet.error.errors[0].message };
+  }
 
   try {
     await prisma.pet.create({
-      data: petDate,
+      data: validatedPet.data,
     });
 
     revalidatePath("/app", "layout");
@@ -19,13 +26,17 @@ export async function AddPet(petDate: PetEssentials) {
   }
 }
 
-export async function GetSinglePet(petId: PetWithId["id"]) {
+export async function GetSinglePet(petId: unknown) {
   await sleeper(2000);
+  const validatedPetId = PetIdSchema.safeParse(petId);
+  if (!validatedPetId.success) {
+    return { error: "Invalid data" };
+  }
 
   try {
     const pet = await prisma.pet.findUnique({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
 
@@ -34,13 +45,17 @@ export async function GetSinglePet(petId: PetWithId["id"]) {
     return { error: "Something went wrong. could not get the Pet." };
   }
 }
-export async function DeletePet(petId: PetWithId["id"]) {
+export async function DeletePet(petId: unknown) {
   await sleeper(2000);
+  const validatedPetId = PetIdSchema.safeParse(petId);
+  if (!validatedPetId.success) {
+    return { error: "Invalid data" };
+  }
 
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
 
@@ -50,18 +65,21 @@ export async function DeletePet(petId: PetWithId["id"]) {
   }
 }
 
-export async function UpdatePet(
-  petId: PetWithId["id"],
-  NewPetData: PetEssentials
-) {
+export async function UpdatePet(petId: unknown, NewPetData: unknown) {
   await sleeper(2000);
+
+  const validatedPetId = PetIdSchema.safeParse(petId);
+  const validatedPet = FormSchema.safeParse(NewPetData);
+  if (!validatedPet.success || !validatedPetId.success) {
+    return { error: "Invalid data" };
+  }
 
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: NewPetData,
+      data: validatedPet.data,
     });
 
     revalidatePath("/app", "layout");
